desugar_b.rs hands its beta-reduce and unwrap_or desugars to desugar_c.rs, a pure text move under the 800-line ceiling
